requirements.txt
tabula.py
requests 